Move rules deletion from Purger to Ruler API.
