+++ pkg/chunk/purger/tenant_deletion_api.go
@@ -13,7 +13,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/thanos-io/thanos/pkg/objstore"
 
-	"github.com/cortexproject/cortex/pkg/ruler/rulestore"
 	"github.com/cortexproject/cortex/pkg/storage/bucket"
 	cortex_tsdb "github.com/cortexproject/cortex/pkg/storage/tsdb"
 	"github.com/cortexproject/cortex/pkg/tenant"
@@ -22,24 +21,22 @@ import (
 
 type TenantDeletionAPI struct {
 	bucketClient objstore.Bucket
-	ruleStore    rulestore.RuleStore
 	logger       log.Logger
 	cfgProvider  bucket.TenantConfigProvider
 }
 
-func NewTenantDeletionAPI(storageCfg cortex_tsdb.BlocksStorageConfig, cfgProvider bucket.TenantConfigProvider, ruleStore rulestore.RuleStore, logger log.Logger, reg prometheus.Registerer) (*TenantDeletionAPI, error) {
+func NewTenantDeletionAPI(storageCfg cortex_tsdb.BlocksStorageConfig, cfgProvider bucket.TenantConfigProvider, logger log.Logger, reg prometheus.Registerer) (*TenantDeletionAPI, error) {
 	bucketClient, err := createBucketClient(storageCfg, logger, reg)
 	if err != nil {
 		return nil, err
 	}
 
-	return newTenantDeletionAPI(bucketClient, cfgProvider, ruleStore, logger), nil
+	return newTenantDeletionAPI(bucketClient, cfgProvider, logger), nil
 }
 
-func newTenantDeletionAPI(bkt objstore.Bucket, cfgProvider bucket.TenantConfigProvider, ruleStore rulestore.RuleStore, logger log.Logger) *TenantDeletionAPI {
+func newTenantDeletionAPI(bkt objstore.Bucket, cfgProvider bucket.TenantConfigProvider, logger log.Logger) *TenantDeletionAPI {
 	return &TenantDeletionAPI{
 		bucketClient: bkt,
-		ruleStore:    ruleStore,
 		cfgProvider:  cfgProvider,
 		logger:       logger,
 	}
@@ -63,38 +60,12 @@ func (api *TenantDeletionAPI) DeleteTenant(w http.ResponseWriter, r *http.Reques
 
 	level.Info(api.logger).Log("msg", "tenant deletion mark in blocks storage created", "user", userID)
 
-	if api.ruleStore != nil {
-		err := api.deleteRules(r.Context(), userID)
-		if err != nil {
-			level.Error(api.logger).Log("msg", "failed to delete tenant rule groups", "user", userID, "err", err)
-			http.Error(w, errors.Wrapf(err, "failed to delete tenant rule groups").Error(), http.StatusInternalServerError)
-			return
-		}
-	}
-
 	w.WriteHeader(http.StatusOK)
 }
 
-func (api *TenantDeletionAPI) deleteRules(ctx context.Context, userID string) error {
-	if !api.ruleStore.SupportsModifications() {
-		level.Warn(api.logger).Log("msg", "cannot delete tenant rule groups, using read-only rule store", "user", userID)
-		return nil
-	}
-
-	err := api.ruleStore.DeleteNamespace(ctx, userID, "") // Empty namespace = delete all rule groups.
-	if err != nil && !errors.Is(err, rulestore.ErrGroupNamespaceNotFound) {
-		return err
-	}
-
-	level.Info(api.logger).Log("msg", "deleted all tenant rule groups", "user", userID)
-	return nil
-}
-
 type DeleteTenantStatusResponse struct {
-	TenantID                  string `json:"tenant_id"`
-	BlocksDeleted             bool   `json:"blocks_deleted"`
-	RuleGroupsDeleted         bool   `json:"rule_groups_deleted"`
-	AlertManagerConfigDeleted bool   `json:"alert_manager_config_deleted,omitempty"` // Not yet supported.
+	TenantID      string `json:"tenant_id"`
+	BlocksDeleted bool   `json:"blocks_deleted"`
 }
 
 func (api *TenantDeletionAPI) DeleteTenantStatus(w http.ResponseWriter, r *http.Request) {
@@ -113,27 +84,7 @@ func (api *TenantDeletionAPI) DeleteTenantStatus(w http.ResponseWriter, r *http.
 		return
 	}
 
-	result.RuleGroupsDeleted, err = api.isRulesForUserDeleted(ctx, userID)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
 	util.WriteJSONResponse(w, result)
-}
-
-func (api *TenantDeletionAPI) isRulesForUserDeleted(ctx context.Context, userID string) (bool, error) {
-	if api.ruleStore == nil {
-		// If API doesn't have access to rule store, then we cannot say that rules have been deleted.
-		return false, nil
-	}
-
-	list, err := api.ruleStore.ListRuleGroupsForUserAndNamespace(ctx, userID, "")
-	if err != nil {
-		return false, errors.Wrap(err, "failed to list rule groups for tenant")
-	}
-
-	return len(list) == 0, nil
 }
 
 func (api *TenantDeletionAPI) isBlocksForUserDeleted(ctx context.Context, userID string) (bool, error) {
