+++ pkg/ruler/store_mock_test.go
@@ -116,10 +116,6 @@ func newMockRuleStore(rules map[string]rulestore.RuleGroupList) *mockRuleStore {
 	}
 }
 
-func (m *mockRuleStore) SupportsModifications() bool {
-	return true
-}
-
 func (m *mockRuleStore) ListAllUsers(_ context.Context) ([]string, error) {
 	m.mtx.Lock()
 	defer m.mtx.Unlock()
